requirements.txt
Django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.2
django-cors-headers>=3.15
psycopg2-binary>=2.9  # لو مستخدم PostgreSQL
python-dotenv>=1.0.0
Pillow>=9.0.0
django-filter>=23.1